pyproject.toml
[tool.poetry]
name = "svelteaction"
version = "0.1.0"
description = "The doc for Svelte Actions"
authors = ["Mohamed-Kaizen <m.n.kaizen@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
mkdocs-material = "^8.5.2"
mkdocs-minify-plugin = "^0.5.0"
mkdocs-git-revision-date-localized-plugin = "^1.1.0"
mkdocs-git-authors-plugin = "^0.6.4"
mkdocs-git-committers-plugin-2 = "^1.1.1"
mkdocs-macros-plugin = "^0.7.0"
Pillow = "^9.2.0"
CairoSVG = "^2.5.2"
PyYAML = "^6.0"


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
darglint = "^1.8.1"
flake8 = "^5.0.4"
flake8-bandit = "^4.1.1"
flake8-black = "^0.3.3"
flake8-bugbear = "^22.9.23"
flake8-docstrings = "^1.6.0"
flake8-import-order = "^0.18.1"
isort = "^5.10.1"
radon = "^5.1.0"
safety = "^2.3.1"
mypy = "^0.982"

[tool.isort]
profile="black"

[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
